BUG#31667659 SHELL DUMP: FETCHING METADATA TRIGGERS/HISTOGRAM/STATUS IS INEFFICIENT
Tasks which are chunking the tables were accumulated in a container,
then pushed all at once. If instance contains large number of tables,
this causes a significant delay.

Push each table task to the worker queue as soon as it's available, do
not wait till all tasks are ready.

Synchronized_queue now handles tasks with priority.

Change-Id: Ie88e2d859eb4a44ddad7e43fb3bffd4ceefb65e6