BUG#31545728: RANDOM CORE DUMP WHEN ADDING AN INSTANCE
Depending on the target instance system characteristics and dataset,
Shell crashed randomly during the .addInstance() operation if clone was
used as the recovery method and the progress was being monitored using
progress bars (default).

The progress monitoring implementation assumes that Performance_Schema
will have the information of the 4 data transfer stages of MySQL clone
whenever it's queried for the first time. However, it's possible that
only 1 or 2 stages are displayed in P_S when Shell first queries it.
This is very common when the dataset is very small.

This assumption caused a segmentation fault, resulting in a crash.

This patch fixes that by ensuring that progress updates are performed
according to what information is available at each time.

Added unit and system tests to verify the patch.

Change-Id: I6cdd250d3d080e500c348cb79ea581b13030875a