BUG#31706755 Cache and aggregate queries executed by dumper
This patch introduces a cache with instance metadata, which stores data
frequently used by the dumper. Queries which fetch most of this data
were rewritten in order to acquire aggregate results, meaning data is
obtained for all the DB objects at once, instead of querying them one by
one. This reduces the overall number of queries executed by the dumper,
improving the performance.

BUG#31706755 MYSQL SHELL DUMP: WHEN FETCHING A CHUNK, ORDER BY PK

Both chunking and dumping queries were ordering the data using the first
column in a unique index (if it was available for the dumped table),
this patch changes this behaviour to use all the columns.

Change-Id: I8286b745dd4069d294d976fc673f1424000c12f2