Bug #31673163	REBOOTCLUSTERFROMCOMPLETEOUTAGE() IS MISSING CHECK FOR PENDING CERT TRANSACTIONS
rebootCluster was only considering GTID_EXECUTED when validating for the
instance that has the most transactions. Transactions that were received
(and certified) but not executed yet must be included in that check.
Also, transactions received through replication channels other
than the GR applier must also be considered in that check, since that
member was probably the PRIMARY while the cluster was running and
ignoring these transactions would lead to data loss. Only known/managed
channels must be considered, however, since unknown channels may have
stale transactions not meant to be applied.

Change-Id: I7b6df3f57928006ec6425d7e3041beaaf4d2b10b