BUG#31798495 IPWHITELIST OPTION WAS DEPRECATED IN GROUP REPLICATION BUT NOT IN ADMINAPI
In 8.0.22, Group Replication deprecates group_replication_ip_whitelist in
favor of group_replication_ip_allowlist. The deprecation must be reflected
in the AdminAPI as well.

This patch fixes that by adding a new option named "ipAllowlist" and by
deprecating "ipWhitelist" in dba.createCluster(), Cluster.addInstance()
and Cluster.rejoinInstance().

Regardless of the target instance version, if ipWhitelist is used Shell
will print a deprecation warning.

Considering that if the target instance is 8.0.22 GR reports a
deprecation warning, the AdminAPI will avoid that by ensuring the right
variable is set according to the version:

If target instance is >= 8.0.22 and ipWhitelist is used, updates the new
system variable. Otherwise, updade the old system variable.

This patch also fixes a bug in cluster.options() when the option "all"
is enabled. The resulting list of options was excluding options as
presented when the option is not enabled.

Added new tests to verify the changes and updated existing tests.

Change-Id: I04b04e2264f104eaa60c933ef498cbfe7a07aa45