BUG#31491092 ERROR WHILE SETTING INNODB CLUSTER WITH SETUPADMINACCOUNT(USER) IN MYSQL 8.0.20
The functions .setupAdminAccount() and .setupRouterAccount() would fail
if the validate_password plugin was enabled with an error indicating the
password did not meet the policy requirements. This happened regardless
if the password  met or not the policy requirements.

This happened because the functions were creating the user in 2 separate
transactions:

1) First: "CREATE USER IF NOT EXISTS...", without setting a password

2) Then, "ALTER USER ... IDENTIFIED BY ...".

The password validation policies, when enabled, don't allow creating a
user without a password.

This patch fixes that by ensuring that the username is created with the
password set.

Added regression tests.

Change-Id: I7dc2abd06e28092f73dd04823e962d008feac22f