BUG#31798724,BUG#31467823 Terminology updates
BUG#31798724 ADMINAPI MESSAGES REFER TO DEPRECATED COMMANDS SUCH AS START/STOP SLAVE

BUG#31467823 REPLACE MASTER/SLAVE TERMINOLOGY WITH PRIMARY/REPLICA

In 8.0.22, MySQL Replication deprecates the following commands:

- START/STOP SLAVE
- SHOW SLAVE STATUS
- SHOW SLAVE HOSTS
- RESET SLAVE

In favor of the new:

- START/STOP REPLICA
- SHOW REPLICA STATUS
- SHOW REPLICAS
- RESET REPLICA

On top of that, all references to "master" were replaced with "source"
and all references to "slave" were replaced with "replica".

However, the AdminAPI has output user messages with references to the
deprecated commands and/or terms and makes use of the old commands
regardless of the target instance version.

This patch fixes that by ensuring the new terminology is used in user
output messages/errors and by ensuring that the new commands are used
when the target version is >= 8.0.22.

Updated existing tests to match the changes.

Change-Id: Iba9fba8687e9f70ca339b07b918a09ee5ec851ce