BUG#31691232 DBA.CONFIGUREINSTANCE() IS NOT POSSIBLE ON A SERVER PART OF A RUNNING GROUP
dba.configureLocalInstance() and dba.configureInstance() were not
allowed to be used against instances that were part of unmanaged
replication groups. This made impossible to create a clusterAdmin
account to be used to adopt the group into an InnoDB cluster and manage
it via Shell.

Creating the clusterAdmin accounts is very important to allow the remote
management of InnoDB clusters and to avoid users to manually deal with
the user creation and required privileges.

This patch fixes that by ensuring dba.configureLocalInstance() and
dba.configureInstance() are allowed to be executed against instances
belonging to unmanaged replication groups. Since the instance will be
ready for Cluster usage, a message indicating that will be presented.

Added regressions tests.

Change-Id: I1314715b82f5589145c48697fc9e40dd10dfca28