BUG#31909408 SHELL'S DUMP UTILITY DIES WITH ERROR "BAD ALLOCATION"
When dumping a table, during the chunking operation, if step becomes too
small, chunker enters an infinite loop, resulting in out of memory
errors. As a fix, when step is too small it is set to lowest possible
valid value.

The accuracy value used by the chunker to estimate when computed range
has enough rows was not computed correctly, this is now fixed.

As a general improvement of the chunking algorithm, if the range that
chunker is currently checking doesn't have enough rows, increase the upper
boundary and try once again. This allows to reduce the number of
zero-sized chunks.

Change-Id: If0c77fb034ca253991d9f80f320ccf2133b5815b