WL#13837, BUG#31975416 Configure parallel applier for replication recovery
Replication applier and incremental recovery currently use a single
thread applier by default. This WL changes that by ensuring that
parallel appliers are enabled by default in InnoDB Clusters and
ReplicaSets, improving the throughput of both the replication applier
and incremental recovery.

This functionality is only supported and enforced for servers running
versions >= 8.0.23.

In order to have parallel appliers enabled by default in InnoDB Clusters
and ReplicaSets, the following set of configurations became a
requimement:

  -  binlog_transaction_dependency_tracking=WRITESET
  -  slave_preserve_commit_order=ON
  -  slave_parallel_type=LOGICAL_CLOCK
  -  transaction_write_set_extraction=XXHASH64

The number of parallel workers is now set to 4 by default, being
possible to change it in:

  - dba.configureInstance()
  - dba.configureReplicaSetInstance()

With a new added option named 'applierWorkerThreads'.

The monitoring functions were extended to include the information about
the parallel appliers options and values:

  - <Cluster>.status() - with 'extended' > 1, it includes the nº of
appliers configured

  - <Cluster>.options() and <ReplicaSet>.options() - include the
settings and values regarding parallel appliers.

Finally, sandboxes using MySQL versions >= 8.0.23 are now deployed
using the new default values that enable parallel appliers.

Added new tests to cover the new features and changes and updated
existing ones to cover all scenarios.

Bugs fixed:

- BUG#31975416 <REPLICASET>.REJOININSTANCE DOESN'T CHECK THE INSTANCE
  CONFIGURATION

<ReplicaSet>.rejoinInstance() was not validating the target instance
configurations prior to attempting a rejoin. It's possible that the user
or some process changes the required configurations meanwhile the
instance is out of the ReplicaSet so the command must re-verify those.

This has been fixed by ensuring that .rejoinInstance() also validates
the required configurations for InnoDB ReplicaSet.

Change-Id: Idbc6010d0bc08c00436ec4101a891b5c42afc33b