Fix for test failures in PB2
PB2 is now using IP ranges on the subnet range of 100.0.0.0/8,
which means ipAllowlist won't be AUTOMATIC in GR. For that reason, the
AdminAPI norecord tests are failing.

This patch fixes that by ensuring the ipAllowlist is set on those tests
to include the IP of the host running the tests.

Change-Id: I3cd46a444ef91c2d953e5eaf920c4706e32cd609