Bug #31907471 SEGMENTATION FAULT ERROR USING SHELL ON DOCKER CONTAINERS
When locale files are missing (/usr/lib/locale/*), which can happen in
certain docker images like Debian, Ubuntu and Fedora; the shell will try
to log an error for being unable to switch it. Because that happens
before the logger is initialized, it would crash.

Also, when mmap is used, an exception would be thrown trying to sync
before anything is written.

Change-Id: I9fcaed178b10ebad7278621fe33a94a06490bc46