BUG#26649039 SHELL FAILS TO IDENTIFY REJOINING MEMBER WITH NEW UUID
An instance within a cluster is identified through it's UUID. If this
changes when the instance gets off the cluster, the instance can be
identified through its host and port.

A new check has been added when an instance is rejoined so when the
instance is not found on the metadata through the UUID, it is searched
using its host and port, if found, the metadata will be updated
based on the options used for the rejoin operation.

This check is executed on rejoinInstance and rescan operations.

Additionally, a check is executed to verify the serverId of all the
members is registered in the metadata as an instance attribute,
if not, the metadata is updated accordingly.

This check is executed on add, rejoin and rescan operations.

Change-Id: I35a56069505d3b2cf83774db061b06e59f7ad525