BUG#31926697 WHEN USING PYTHON 3.8, IT IS POSSIBLE THAT FORMAT STRINGS WILL NOT WORK
Python 3.8 has added a new field to the PyCompilerFlags structure:
`cf_feature_version`. This structure is used in our code, however the
new field was not initialized, meaning it was set to some random value.
If this random value was low enough, an exception was thrown whenever a
Python's format string was used. This patch fixes this problem and
initializes the `cf_feature_version` field with a proper value.

Change-Id: I1950f978357f9cefbbc49921b481aaa78acacd08