Fix output of dump/load methods
Output of dump/load could be hard to read, as sometimes the progress
information is in the same line as status messages.

The mechanism which is used to prevent this behaviour was failing, as
the variable which handles hiding/showing the progress information was
unnamed, and its destructor, which is supposed to show the progress,
was called immediately, instead of at the end of the scope. These
variables are now named, and the issue is fixed.

Change-Id: I23fc7fa8d56ab0b94f242e27f23a47cd2c155a28