Minimum set of privileges required to run dump_instance()
Added a test to verify the minimum set of privileges required to run
dump_instance(), if this test fails, the list needs to be reviewed and
the online docs should be updated.

Added a verification to check if user has the TRIGGER privilege when
requesting to include triggers in the dump.

Change-Id: I3ff0f6b764cabfefa262caad5ee2ba4625cb7da0