Fix test which was failing randomly
When testing load of a partial dump, half of the data files were copied
first, while the rest was copied later. If the first half contained a
data file, while the corresponding IDX file was in the second half, test
would randomly fail. Test was fixed, all IDX files are copied before any
of the data files.

Change-Id: Ifb38344a7c4da322ab5753240901bd48b9aaa80b