BUG#32072961 TRANSACTION_BUFFER::FIND_LAST_ROW_BOUNDARY_BEFORE SEGFAULT
If the dumped data file has to be subchunked when loading, the
Transaction_buffer::find_last_row_boundary_before() method tries to find
the row boundary using the indexes from .idx file, but it was not
checking if the given index does not exceed the size of data in the
buffer, which may result in segmentation fault.

This has been fixed, indexes which point beyond the data in the read
buffer are ignored.

Change-Id: I3d4430257071e2e8316835b10de0c1ffdb21afb9