Normalizing type errors.
There is an inconsistency on how data type errors are handled, in some
cases they are reported as ArgumentErrors while in others they are
reported as TypeErrors.

This patch fixes this inconsistency in a way that:

TypeError is used when the type of the provided argument does not match
the type of the parameter/option.

ArgumentError should be used when the data type matches but for some
reason the value is not valid.

Change-Id: I52d69a57f7d78ced611ac8bd621be3ec67509771