Introducing Option Packs - BUG#31991095 Included Fix
An option pack is a container for options to be used on the exposed API
functions instead of the currently used Dictionary.

The most important advantage of an Option Packs over Dictionaries is it's
ability to automatically generate/store metadata for the supported options:
name, type and whether it is optional or mandatory.

Another difference is that with Option Packs the conversion of the provided
options to the option pack is done transparently BEFORE the target API is
actually called, removing the burden of having to implement unpack logic on the
API functions.

Option Packs support inheritance and aggregation, which enables easily creating
complex structures that adjust to the different API needs.

This refactoring includes switching from Dictionary to Option Packs on the
APIs exposed by the following objects:

- util: All of the functions except importJson (to be done later)
- dba: just deploySandbox instance as a try-out (rest is to be implemented)

NOTE: The only side effect of option packs is that errors reported on option
      reading is reported pre-pending the Argument number with the failure.

BUG#31991095 - BAD ERROR MESSAGE USING BYTESPERCHUNK ON UTIL.IMPORTABLE FUNCTION

The bytesPerChunk option is not valid when loading from multiple files
but that was not clear on the error message, it has been updated to:

"The 'bytesPerChunk' option is invalid when loading from multiple files.

Change-Id: Ia2b939f2822596992ea5ea8d72d7c350bb146218