BUG#31748786 SHELL LOADDUMP SKIPBINLOG=TRUE NOT FAILING ON LACK OF PRIVILEGE
When loading a dump into MDS, the 'skipBinlog' option cannot be used, as
it is not possible to change the session value of the 'sql_log_bin'
server variable there, Shell will now throw an exception in such case.

Added a test to ensure that if user is missing the required privileges
to set the `sql_log_bin` variable, an exception is thrown.

Change-Id: I0d88368bc915b1b27f1c54e87e8a6d98ba52a51d