BUG#32140970 BUG#32165101 util.dumpTables() fixes
BUG#32140970 UTIL.DUMPTABLES() SHOULD SUPPORT OCIMDS OPTION

The util.dumpTables() utility did not support options which would allow
to fix the incompatibilities with MDS, or to load the dump using PARs.

It now supports the following options:
 - 'ocimds'
 - 'compatibility'
 - 'ociParManifest'
 - 'ociParExpireTime'
These options behave exactly the same as corresponding options in the
other dump utilities.

The 'ignoreVersion' option can now be used in util.loadDump() in order
to load a dump created without the 'ocimds' option into MDS. Previously,
loading such dump would result in an exception.

BUG#32165101 DUMPTABLES() SHOULD BEHAVE AS DUMPSCHEMAS() WITH A SINGLE SCHEMA

When using util.dumpTables(), the dump was created without the SQL to
recreate the schema containing tables/views being dumped, so the dump
had to be loaded into an existing schema, either the active one in the
global shell session, or the one specified by the 'schema' option of the
util.loadDump() function.

This behaviour has been changed, dumps will now contain the schema SQL
statements. When loading the dump, loadDump() will by default recreate
the schema (this is consistent with loading the dumps created by other
dump utilities). It is still possible to load the dump into another
existing schema using the 'schema' option.

The dumps created by the previous version of the util.dumpTables() need
to be loaded using the 'schema' option, the active schema of the global
shell session is no longer used as the target schema if 'schema' is not
given.

Change-Id: I6f8b820aed5b904c7a0f0df12a67e375a10bd8cc