In ReplicaSet, whenever replication stops due to an error such as the one caused by a diverging transaction in a replica, the applier thread stops.
On those scenarios, when changing the replication configuration with
`CHANGE MASTER` an error will be thrown indicating the command cannot be
executed since the slave was stopped due to an error. This is a
protection mechanism to avoid possible gaps in the multi-threaded
applier.

In ReplicaSet, this can happen in .rejoinInstance() since the command
resets replication by doing:

- STOP REPLICA
- CHANGE MASTER

To overcome that, and to allow users to rejoin instances back to the
cluster after the replication errors are solved, the command must also
reset the channel.

This patch fixes that by ensuring a `RESET REPLICA` is performed too
when resetting the replication channel in .rejoinInstance().

Change-Id: I0ddabfa078ff521c1321a882e5a3455b04622614