BUG#31428813 : DBA.UPGRADEMETADATA() FAILS WITH ERROR: UNKNOWN COLUMN 'MYSQL ROUTER' IN 'FIELD LIST'
Upgrading the Metadata schema with dba.upgradeMetadata() failed if
ANSI_QUOTES were used as sql_mode. The command would fail with the
following error: ERROR: Unknown column 'MySQL Router' in 'field list'.

This was caused by a specific query to insert data into the routers
table of the metadata schema which was using double quotes to quote a
string. When the sql_mode is set to use ANSI_QUOTES, MySQL treats " as
an identifier quote character and not as a string quote character
resulting in an error whilst running that query.

This patch fixes that by ensuring the query uses single quotes to quote
strings. This is the ANSI rule and recommendation for SQL: use single
quotes to identify strings.

On top of that, this patch ensures that the upgrade metadata command
prepares the session to be used by the AdminAPI, which, amongst other
sanity checks it ensures the sql_mode used by that session uses the
default values to avoid user-set incompatible settings. Additionally,
the same was done for:

 - dba.getCluster()
 - dba.dropMetadataSchema()

Extended upgrade metadata tests to verify that using ANSI_QUOTES and
other sql_modes do not break the upgraded.

Change-Id: Ib436a6ed82a58121c1f01f5147e7fa6c9840aee7