BUG#32112864 : REBOOTCLUSTERFROMCOMPLETEOUTAGE() DOES NOT EXCLUDE INSTANCES IF IN OPTION "REMOVEINSTANCES" LIST
dba.rebootClusterFromCompleteOutage() requires an active session on
MySQL Shell to:

  - Obtain the cluster members as registered in the Metadata.
  - Determine which of the cluster members has the GTID superset.
  - If the active session is not to the member with the GTID
    superset the command aborts and indicates to the user which is
    the instance with the GTID superset.

However, the GTID superset check is done using all instances (registered
in the Cluster's metadata) reachable by Shell.

If an instance has a diverging GTID-set and the user wants to explicitly
remove it from the cluster, the operation will be blocked since the
command cannot determine which instance has the GTID superset. Depending
on the point-of-view, different instances can be considered as the most
up-to-date.

Also, it should be possible for users to reboot a cluster by picking a
particular instance even though it's not the one most up-to-date, as
long as they indicate the other instances are not meant to be rejoined
by using the command's options/prompts.

This patch fixes that by ensuring that if the user explicitly sets the
`removeInstances` variable or answers "NO" to the prompt regarding
instances rejoin, those instances must be excluded from the GTID
superset validation.

Added regression tests.

Change-Id: I53ded77e8211902c352880f2fddd67c471b8260b