BUG#32115948 UTIL.LOADDUMP THROWS ERROR WHEN THE DUMP CONTAINS USERS CREATED WITH EXTERNAL AU
Certain authentication plugins are not supported in MDS, trying to load
a dump which contained accounts using such plugins resulted in failure.

The 'ocimds' option of the dump utilities now includes additional checks
for accounts using unsupported authentication plugins, if such account
is detected, it is listed as incompatible. This issue needs to be fixed
before continuing the dump. In order to automatically remove invalid
accounts, a new compatibility option has been added:
  'skip_invalid_accounts'.

Currently, the following authentication plugins are supported:
 - caching_sha2_password
 - mysql_native_password
 - sha256_password

The version of the dump utility has been bumped to 1.0.2, loading a dump
with an older version will result in a notice to recreate the dump in
case of any errors.

Change-Id: Ia45b0e31ec65e013e3be68696da2d9c3c59f4a76