BUG#32152133 : REPLACE CHANGE MASTER/START SLAVE TERMINOLOGY WITH CHANGE SOURCE/START REPLICA
In 8.0.23, MySQL Replication deprecated the following commands:

- CHANGE MASTER TO
- START SLAVE UNTIL MASTER_LOG_POS, MASTER_LOG_FILE

In favor of the new:

- CHANGE REPLICATION SOURCE TO
- START REPLICA UNTIL SOURCE_LOG_POS, SOURCE_LOG_FILE

And their parameters:

- MASTER_HOST
- MASTER_PORT
- MASTER_*

In favor of:

- SOURCE_HOST
- SOURCE_PORT
- SOURCE_*

This patch fixes that by ensuring the new terminology is used in log
messages/errors and by ensuring that the new commands are used when
the target version is >= 8.0.23.

Updated existing tests to match the changes.

Change-Id: I6559a243b10f4e081b32a2c23e5f535244a519f5