BUG#32197197 : ADMINAPI DOES NOT PROPERLY CHECK FOR PRECONFIGURED REPLICATION CHANNELS
With BUG#29305551, dba.checkInstanceConfiguration() was extended to
include a check to verify if asynchronous replication is configured and
running on the target instance and print a warning if that's the case.
On top of that, the same check is used in <Cluster>.addInstance() and
<Cluster>.rejoinInstance() to terminate the commands with an error if
such scenario is verified.
The same check is also used in dba.rebootClusterFromCompleteOutage()
whenever there are instances to be rejoined to the cluster.

However, the check missed verifying if the replication channel was
configured but NOT running. This must be verified as well to completely
ensure that no replication channels are set and possibly started later
on an InnoDB Cluster.

This patch fixes that by ensuring the verification also takes in
consideration replication channels configured but NOT actively running.
On top of that, an erroneous message indicating the possibility of using
'STOP REPLICA' to override this check was removed and replaced with an
informative message clearly indicating that unmanaged replication
channels are not supported and possible dangers of its usage.

Added regression tests to verify the new behavior.

Change-Id: Ib7bc9c4b6898bb44b01de166b34c50dc0f27f7f5