BUG#32197222 : ADMINAPI CREATECLUSTER() SHOULD NOT ALLOW EXISTING ASYNC REPLICATION CHANNELS
With BUG#29305551, dba.checkInstanceConfiguration() was extended to
include a check to verify if asynchronous replication is configured and
running on the target instance and print a warning if that's the case.
On top of that, the same check is used in <Cluster>.addInstance() and
<Cluster>.rejoinInstance() to terminate the commands with an error if
such scenario is verified.
The same check is also used in dba.rebootClusterFromCompleteOutage()
whenever there are instances to be rejoined to the cluster.

However, dba.createCluster() was erroneously skipping that verification.
The same for dba.rebootClusterFromCompleteOutage() which was skipping
that verification on the instance being used to bootstrap the group.

This patch fixes that by ensuring that the check is also performed
whenever creating or rebooting a cluster from complete outage.

On top of that, it adds support to override that verification on
dba.createCluster() by making use of the 'force' option, and it improves
the error messages.

Added regression tests.

Change-Id: I2be44f52835238407685ee2f1d2918f77efaf0ce