BUG#31882876 : MYSQL ROUTERS MUST BE REBOOTED TO PROCEED UPGRADE METADATA
When upgrading the metadata using dba.upgradeMetadata(), if there are
Router instances that need to be upgraded, the command waits until
all Router instances are upgraded before continuing. The command offers
the users an option to re-check for outdated Routers and continue with
the metadata upgrade.

Router upgrade is only complete after a restart of it, however, the
message printed in the command didn't mention that.

This patch fixes that by extending that message to include the
information that the Router instances must be restarted after their
binaries are upgraded.

Updated existing tests to meet the changes.

Change-Id: I1252fb0c959a820effb969b9090e209da89be31d