Fixing Paramiko/OCI-SDK Loading problems.
In cryptograpgy 3.2.1 an exception is raised when OpenSSL < 1.1.0 is used.
This causes Paramiko/OCI-SDK to not be usable in platforms having a lower
version of OpenSSL in the system.

This patch defines environment variable CRYPTOGRAPHY_ALLOW_OPENSSL_102
to turn the error into a warning and keep the Paramiko/OCI-SDK functionality
available in such platforms.

Change-Id: I7c07c0b0718bfaf80ac52c936abf5c54d58a7077