Fix compilation with Python 3.8 and clang
In Python 3.8, tp_print slot has been marked as deprecated, since it's
initialized in our code in order to avoid missing initializer warning,
it generated the deprecated warning instead. As all warnings are treated
as errors, compilation fails in such case.

The deprecation warning has been disabled for the parts of the code
which initialize the tp_print slot.

Change-Id: Id35797b82e01879f09d6dfa5812a62475d0a6e57