BUG#32015164: MISSING INFORMATION ABOUT REQUIRED PARALLEL-APPLIERS SETTINGS ON UPGRADE SCNARIO
In 8.0.23, having parallel-appliers enabled by default became a
requirement for InnoDB Cluster/ReplicaSet.

However, when upgrading a cluster/replicaset from a version previous to
8.0.23 to a version >= 8.0.23, those settings won't be in place. This
doesn't break the cluster/replication, but by design we must ensure
parallel-appliers are enabled. The user must be warned about this and
guided on how to fix it.

This patch fixes that by adding a new verification to check if the
parallel-applier settings are not enabled on cluster/replicaSet members
running 8.0.23+ and includes that information as a NOTE in the
'instanceErrors' attribute of the output of .status().
A note indicating to use
dba.configureInstance()/dba.configureReplicaSetInstance() is also
included in the message.

Added regression tests to verify the changes.

Change-Id: I65dc1f58c4f0a5e202ba6d65681f4409286110cb