BUG#32213605 LIST OF PRIVILEGES ALLOWED IN MDS IS OUTDATED
List of privileges allowed in MDS was outdated, if there was an account
which had any of the missing privileges, and util.dumpInstance() was
called with both `users` and `ocimds` both set to true, compatibility
errors were reported.

Missing privileges have been added to the list:
 - REPLICATION_APPLIER
 - APPLICATION_PASSWORD_ADMIN
 - CONNECTION_ADMIN
 - RESOURCE_GROUP_ADMIN
 - RESOURCE_GROUP_USER
 - XA_RECOVER_ADMIN

The USAGE privilege has been also reported as disallowed, this is now
fixed as well.

Change-Id: I9e540a75c714ba65bd6e0d58760e8beaf77a1e8c