WL#14358 Log SQL statements via syslog
New command line option

  mysqlsh --syslog

Corresponding run-time option (Boolean):

  history.sql.syslog

Specifies whether the following should be logged to the system log
(syslog on Unix and Windows Event Log on Windows):
 - SQL statements executed in SQL mode
 - \source commands executed in SQL mode
 - SQL statements executed using the \sql command

These statements are logged only if Shell is running in interactive mode.

The log message contains the following information (in order):
 - SYSTEM_USER - The operating system user name.
 - MYSQL_USER - The MySQL user name.
 - CONNECTION_ID - The client connection identifier.
 - DB_SERVER - The server host.
 - DB - The default database.
 - QUERY - The text of the logged statement.

The log message is formatted as a series of <key>=<value> pairs
separated with the space character, values can be optionally quoted
with the ' character.

The log message is truncated to 1024 bytes.

If the SQL statement to be executed would match the filters specified
using the --histignore command line option (or its corresponding
run-time option shell.options["history.sql.ignorePattern"]), it is not
going to be logged.

Change-Id: Id0cbfbac99914ddafa7af115ed1f2a0a0d3ed995