WL#14297 BUG#31186637 BUG#31829337 - Shell Command Line Integration Improvements
Metadata Based CLI
==================
This patch enhances the CLI for APIs by using the function and option metadata
to properly parse the command line arguments provided by the user and associate
them to the corresponding parameter on the API call.

The new logic relies on the fact that metadata information is available for both
parameters as well as for options in dictionary parameters, which is achieved by
using Option Packs rather than dictionaries.

CLI Support For Plugins
=======================
A new boolean option "cli" is introduced for the definition parameter at the
shell.addExtensionObjectMember function. When a function is registered setting
the cli option to TRUE, the object and the functions with CLI enabled will be
available for CLI calls.

This patch adds support for:
- List arguments: positional argument with a comma separated list of values
- List options: named argument with a list of values which can be defined
            several times, aggregating the values on the target option.
- Handling of conflicting options: same option valid for more than one parameter.
- Short name for options.
- Support for user defined data types: --option[:type]=value
- Support for help retrieval in CLI format: provides information about the syntax
and the arguments and options allowed on the command:

- $ mysqlsh -- object --help/-h
- $ mysqlsh -- object function --help/-h

Bug #31186637
=============
SETOPTION TAKES INT VALUES AS STRINGS

The CLI integration was processing all the arguments in the CLI as strings, causing
the error described on the bug. The CLI integration has been improved in a way that
the data is interpreted based on what the API functions expect to receive.

There are still some cases where the function may handle different data types.

If that is the case of an anonymous (positional) parameter, (i.e. the 'value'
parameter in the set-option cluster function) the parameter will be interpreted
as follows:

- Valid numeric sequences will be interpreted as numbers
- Valid JSON document/array will be interpreted as a Dictionary or List
- true/false will be interpreted as booleans
- Rest will be interpreted as strings.

In the case of a named parameter, interpretation will be done the same way except
that the user has the ability to force a data type by defining the target type as
follows:

--<option>[:<type>]=<value>

Valid types include: str, bool, int, uint, float, json, bool, null

Bug #31829337
=============
IMPROVE USABILITY OF DUMP OPERATIONS IN CLI: LIST PARAMETERS NOT SUPPORTED

This enhancement introduces support for lists in CLI calls, including:

- List anonymous (positional) parameters
- List named parameters

When a positional argument requires a list of values there are 2 ways to specify
the items:

- As a comma separated list of values will be split to create the items.
- Any additional anonymous argument will be considered a list item.

(because of the second rule above, NONE of the shell APIs contain positional
arguments after a positional list argument)

When a named argument requires a list of values there are 2 ways to specify the
items:

- As a comma separated list of values: --option=value1,value2,value3,...
- Calling the same option repeated times: --option=value1 --option=value2

Change-Id: If4295c54b5eb2de5de115db58b016febecfc66e2