BUG#32369064 LOGIN PATH CREDENTIAL HELPER NOT WORKING
The login-path helper stopped to work correctly after the fix for
BUG#19953349 has been applied in the server (mysql_config_editor) code.
The fix has changed the way the user, host, socket and password values
are stored in the login file: they are now quoted with the " character,
while any " character within the value is escaped with the backslash
character.

The login-path helper has been fixed as follows:
- When storing the password, helper was already quoting the value in
  order to overcome the issue reported in BUG#19953349, hence this behavior
  remains unchanged if mysql_config_editor < 8.0.24 is used. If a newer
  version is used, password is not quoted, however the  backslash
  characters within its value are escaped in order to fix another issue
  present in mysql_config_editor: if a value contains a valid escape
  sequence (backslash + another character, i.e. "\n") it is stored as
  is, but when it's read, the escape sequences are replaced with
  corresponding characters (i.e. with a newline character).
  This fix means that helper will store the same sequence of characters,
  regardless of mysql_config_editor used.
- When reading the non-password entries from the login file their values
  are now unquoted. If they are not quoted (pre 8.0.24) they remain
  unchanged.

Helper was and still is using my_load_defaults() function to read the
password value, the fix for BUG#19953349 did not change anything in that
function, hence this part of the code remains unchanged.

Change-Id: I368d44ddd716cc3343bf3a9129caaf171ee09ca9