BUG#32426083 BUG#28542904 BUG#29853828 Adding CLI support for dba functions
BUG#32426083 CLI HELP FOR DBA FUNCTIONS IS MISSING OPTION DETAIL

This patch replaces the Dictionary options in the DBA functions for
Option Packs. The result of this is full CLI support for functions
in the DBA object, including the CLI help integration.

BUG#32157120 ADMINAPI BUILT-IN HELPER UNNECESSARILY DESCRIBES COMMANDS
             EXCEPTIONS

Partially fixed as this patch removes the exception information for the
functions in the DBA Object.

BUG#28542904 EXCEPTION NOT THROWN FOR DBA.CONFIGUREINSTANCE() DESPITE
             HELP SAYS SO

It has been decided that the exception information should not be part
of the API function help (Details at BUG#32157120)

BUG#29853828 DBA.STOPSANDBOXINSTANCE(): ARGUMENTERROR IF THE ROOT
             PASSWORD IS MISSING ON THE

It has been decided that the exception information should not be part
of the API function help (Details at BUG#32157120)

Change-Id: I3317de3b2f9439128faaf3e7787bb317798462f8