BUG#32850177 FAILED DUMP OPERATIONS LACK CONTEXT
When a dump operation failed due to a server error, the error message
lacked a context, it was difficult to tell what caused the error.

The error messages have been improved, the step which failed is now
included in the error message, as well as schema/table which was being
processed.

Additional fixes and improvements for the dump utilities:

 - Cache frequently used values, i.e. quoted table names.

 - When selecting a key to use during chunking, if table does not have
   a PK, choose index which has less columns, prefer integer columns,
   avoid using NULL columns.

 - When chunking a table, display all columns from the index used.

 - When chunking a table using a non-integer column, use the whole
   composite key, reduce the number of used queries by half.

 - When chunking a table using an integer column, if data range is close
   to the number of rows, use constant step.

Change-Id: I5a82aa5e1184c7503bbc0e2742100c3a9030e692