Minor improvements for util.loadDump()
If opening an OCI object fails, full path is reported in the error
message instead of just the name.

When setting the connection options for sessions created during the
load, max_allowed_packet client variable is set to 1GB unless specified
explicitly by the user (just like in case of the dump counterpart).

Change-Id: I653c7fe43dc0d83e9ae048d2cdf2e550292aeaf0