BUG#30516645 MAKE SURE CLIENT CONNECTIONS USE THE CORRECT COLLATIONS
The characters sets and collations used by sessions to 5.7 servers were
not consistent and not compatible with the UTF-8 string representation
used by Shell.

The collation used by X sessions to 8.0 servers was not consistent,
which could cause problems when running queries with character set
conversions. In some versions of the 8.0 servers collation_connection
was erroneously set to 'utf8mb4_general_ci'.

This issue is now fixed, after a classic connection is established to
a 5.7 server, the `SET NAMES 'utf8mb4'` statement is executed.

When a X session connection is established to the affected 8.0 server,
the 'collation_connection' session variable is set to
'utf8mb4_0900_ai_ci'.

Change-Id: I579226b46ce9d89dde311e553cd437339d337be0