BUG#32376447 MYSQL SHELL DUMPSCHEMA ERROR
When using the dump utilities with 5.6 server, if table data was
chunked, dump failed with "get_uint(9): field type is String" error.

The output of EXPLAIN SELECT statement used when chunking the data is
different in 5.6 server, the "rows" column is at different offset.

This issue is now fixed, when chunking tables of 5.6 server the correct
index is used.

Additionally, the server version information is now cached when using
the classic protocol, speeding up subsequent access.

Change-Id: I273738fd800fee2f0fe38ce9a67d4f1fd0dabdc0