BUG#32381513 INCLUDE MEMBERROLE IN CLUSTER.STATUS({EXTENDED:0})
The memberRole is now included in the default output of cluster.status() (before it was only included when extended had a value >= 1). And the reason behind that is that it's very valuable for users to know what's the role of an instance regardless of its mode (R/W or R/O). For example, it's possible to make a primary to be shown as having the role R/O by simply enabling super_read_only on it.)

Change-Id: Id1181eb886c84e12c85fd6b7849dd05f0eb8e90c