WL#11894 : AdminAPI - ClusterSet.status() and ClusterSet.describe()
This patch introduces a function to monitor the status of a ClusterSet
and its member Clusters:

  - <ClusterSet>.status()

The function gathers information from each member of the ClusterSet and
the replication channel to produce a thorough (depending on the selected
verbosity) status of the ClusterSet as a whole.

It also introduces a function to describe the structure of a ClusterSet:

  - <ClusterSet>.describe()

On top of that, it adds support for the 'manualStartOnBoot' option in
<ClusterSet>.createReplicaCluster() and ensures the value is set in
<Cluster>.addInstance() according to the value defined at Cluster
creation time.

Change-Id: I5229b68075104fd1165638b73c4f7c002dd5f6f6